+++ string.h
@@ -37,6 +37,6 @@ extern char **str_split(char *s, char delim, int *parts);
 extern double str_real(const char *s, int *error);
 extern long long str_long(const char *s, int *error);
 extern unsigned long long str_ulong(const char *s, int *error);
-extern int str_from_sid(char *dest, long sid);
-extern long str_to_sid(char *str);
+extern int str_from_sid(char *dest, long long sid);
+extern long long str_to_sid(char *str);
 extern char *str_urldecode(char *src);
